+++ src/features/ProductListing/ProductListing.tsx
@@ -2,10 +2,11 @@ import { ProductCard } from "@/components/ProductCard"
 import { Product } from "@/shared/types/global.types"
 
 interface ProductListingProps {
-  products: Product[]
+  products: Product[];
+  handleProductClick: (product: Product) => void;
 }
 
-export const ProductListing = ({ products }: ProductListingProps) => {
+export const ProductListing = ({ products, handleProductClick }: ProductListingProps) => {
   if (products.length === 0) {
     return (
       <div>No products available</div>
@@ -15,7 +16,7 @@ export const ProductListing = ({ products }: ProductListingProps) => {
   return (
     <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
       { products.map((product) => (
-        <ProductCard key={product.name} product={product} />
+        <ProductCard key={product.name} product={product} handleProductClick={handleProductClick} />
       ))}
     </div>
   )