+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.6.0 (2026-05-08)
+
+### Pull Requests
+[#7](https://github.com/RafaelMoro/fe-tehesa/pull/7) | Add Drawer to show variations selection on a product
+
+
+## v0.5.0 (2026-05-08)
+
+### Pull Requests
+[#6](https://github.com/RafaelMoro/fe-tehesa/pull/6) | Add Price and variant info to Product | Improve card design
+
+
 ## v0.4.0 (2026-05-05)
 
 ### Pull Requests